Only redefine `require` in top-level filter file
Whenever `require` is called from within files which themselves are
`require`d, the working directory has already been changed to the
filter's directory. Don't change the working directory again in this
case. This led to errors when the filter was called from a subdirectory,
because then the filter's directory would be `../..` or similar, and
every subsequent call to `require` would change the working directory
another two levels up, at which point the `require`d files couldn't be
found anymore. I think the reason this didn't lead to errors previously
is that I only ever called this filter by specifying an absolute path.